+++ src/index.ts
@@ -14,30 +14,27 @@
 import { aligned_alloc_type, pv_free_type, buildWasm } from "./wasm";
 
 import {
-  DB_NAME,
-  DB_VERSION,
-  STORE_NAME,
-  PV_FILE_STORE,
   arrayBufferToBase64AtIndex,
   arrayBufferToStringAtIndex,
   base64ToUint8Array,
   fetchWithTimeout,
-  getDB,
-  getPvStorage,
+  fromBase64,
+  fromPublicDirectory,
   isAccessKeyValid,
-  stringHeaderToObject
+  stringHeaderToObject,
 } from './utils';
 
 import {
   PvFile,
-  fromBase64,
-  fromPublicDirectory
+  getDB,
+  DB_NAME,
+  DB_VERSION,
+  PV_FILE_STORE
 } from "./pv_file";
 
 const dbConfig = {
   DB_NAME,
   DB_VERSION,
-  STORE_NAME,
   PV_FILE_STORE
 };
 
@@ -47,17 +44,16 @@ export {
   pv_free_type,
   buildWasm,
   // utils exports
-  dbConfig,
   arrayBufferToBase64AtIndex,
   arrayBufferToStringAtIndex,
   base64ToUint8Array,
   fetchWithTimeout,
-  getDB,
-  getPvStorage,
   isAccessKeyValid,
   stringHeaderToObject,
-  // PvFile
-  PvFile,
   fromBase64,
-  fromPublicDirectory
+  fromPublicDirectory,
+  // PvFile
+  dbConfig,
+  PvFile,
+  getDB,
 };
